Set mobile session and add mobile content-type
This way you don't have to always have mobile=x in the url. Just do it
once, and then you can unmobilize by setting it back to 0.

The mobile content-type is inert. It doesn't actually do anything, but
it puts it in place so that you can respond with different formats in
`respond_to` if you choose to.